解决菜单排序长度不够用的问题

+++ bachelor-iam/bachelor-acm/bachelor-acm-biz/src/main/java/cn/org/bachelor/iam/acm/domain/Menu.java
@@ -63,7 +63,7 @@ public class Menu {
      * 排序
      */
     @Column(name = "SEQ_ORDER")
-    private Short seqOrder;
+    private Long seqOrder;
 
     /**
      * 更新时间
